feat: Move query splitting into the tokenizer workspace
